backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
motor
google-generativeai
python-dotenv
pydantic
python-multipart
python-dateutil
dnspython
nltk
textblob
scikit-learn
structlog
pytest
pytest-asyncio
python-jose[cryptography]
PyJWT
httpx
